test: add tests from #195
