+++ packages/core/__tests__/mocks/index.ts
@@ -1,0 +1,2 @@
+export * from './mock-readable.js';
+export * from './mock-writable.js';
