+++ packages/core/package.json
@@ -47,15 +47,13 @@
   "packageManager": "pnpm@8.6.12",
   "scripts": {
     "build": "unbuild",
-    "prepack": "pnpm build",
-    "test": "vitest run"
+    "prepack": "pnpm build"
   },
   "dependencies": {
     "picocolors": "^1.0.0",
     "sisteransi": "^1.0.5"
   },
   "devDependencies": {
-    "vitest": "^1.6.0",
     "wrap-ansi": "^8.1.0"
   }
 }
